Move code to a separate file and fix some tests
